finishing module and scripts
